Fully implemented the RandomCatImage module!
Well, almost. There are still error responses for each of the commands that I haven't bothered checking for, because it requires first sending an erronous request and getting the error response schema, and then checking if the request returned an error. A bunch of work that won't be the end of the world if I don't get set up before moving on to other modules.